tt_um_jmack2201/test/requirements.txt
pytest==8.2.2
cocotb==1.8.1
Pillow==9.0.1